backend/iris_api/requirements.txt
pandas
django-webpack-loader
djangorestframework
django-cors-headers